refactor into model
