+++ src/models/focus.ts
@@ -1,0 +1,19 @@
+export type Focus = {
+  name: string;
+  description: string;
+};
+
+export const focuses: Focus[] = [
+  {
+    name: 'core',
+    description: 'UCOP & Ellucid Policies (ppm, ppsm, delegations)',
+  },
+  {
+    name: 'apm',
+    description: 'Academic Personnel Manual',
+  },
+  {
+    name: 'unions',
+    description: 'UCOP Union Contracts',
+  },
+];
